[Debug] Add mod code

+++ scode/code_test.cpp
@@ -4,7 +4,7 @@
 #include "str_int.h"
 
 using namespace std;
-//Code test;
+
 
 int main(){
 
@@ -13,9 +13,8 @@ int main(){
     vstr.push_back("lbdd");
     vstr.push_back("bcb");
     vstr.push_back("bc");
-    // = ("aab", "bdd", "bcb");
-    //vector<int> ack = test.toCode(vstr);
-    vector<int> ack = toCode(vstr);
+
+    vector<double> ack = toCode(vstr);
     
     for (int i = 0; i < ack.size(); i++)
         cout << i << ": " << vstr[i] << " -> " << ack[i] << endl;
